+++ packages/react-magma-dom/src/components/Table/TableBody.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react';
+
+export interface TableBodyProps extends React.HTMLAttributes<HTMLElement> {
+  ref?: any;
+  testId?: string;
+}
+
+export const TableBody: React.FunctionComponent<TableBodyProps> = React.forwardRef(
+  ({ children, testId, ...other }: TableBodyProps, ref: any) => {
+    return (
+      <tbody {...other} ref={ref} data-testid={testId}>
+        {children}
+      </tbody>
+    );
+  }
+);